+++ CSD2b/eindopdracht/lowpass.cpp
@@ -6,15 +6,15 @@ Lowpass::Lowpass(){
 }
 Lowpass::~Lowpass(){}
 
-float Lowpass::tick(inputSample){
+double Lowpass::tick(inputSample){
   outputSample = (inputSample * mix->getA()) + (outputSample * mix->getB());
   return outputSample;
 }
 
-void Lowpass::setFeedback(float feedback){
+void Lowpass::setFeedback(double feedback){
   this->feedback = feedback;
   mix->setX(feedback);
 }
-float Lowpas::getFeedback(){
+double Lowpas::getFeedback(){
   return feedback;
 }
